consumer/requirements.txt
confluent-kafka==2.0.2
redis==4.5.1
psycopg2-binary==2.9.5
